🚚[Rename] 훅 & 컨테이너 최종 분리
