+++ src/features/club-create/hooks/index.ts
@@ -1,3 +1,4 @@
 export { useSelectAddress } from './useSelectAddress';
 export { useCreateBookClub } from './useCreateBookClub';
 export { useBookClubForm } from './useBookClubForm';
+export { useImageField } from './useImageField';
